+++ mycodo/databases/models/timer.py
@@ -9,7 +9,9 @@ class Timer(CRUDMixin, db.Model):
     id = db.Column(db.Integer, unique=True, primary_key=True)
     name = db.Column(db.Text, default='Timer')
     is_activated = db.Column(db.Boolean, default=False)
+    timer_type_main = db.Column(db.Text, default=None)
     timer_type = db.Column(db.Text, default=None)
+    method_id = db.Column(db.Integer, default=None)  # TODO: Add ForeignKey in next major release
     relay_id = db.Column(db.Integer, db.ForeignKey('relay.id'), default=None)
     state = db.Column(db.Text, default=None)  # 'on' or 'off'
     time_start = db.Column(db.Text, default=None)