+++ mycodo/flaskforms.py
@@ -1061,12 +1061,28 @@ class SettingsGeneral(FlaskForm):
 # Timers
 #
 
-class Timer(FlaskForm):
-    timer_id = IntegerField('Timer ID', widget=widgets.HiddenInput())
+class TimerAddBase(FlaskForm):
+    timer_type_output = SelectField(
+        lazy_gettext(u'Timer Type'),
+        choices=[
+            ('relay', lazy_gettext(u'Relay (ON/OFF)')),
+            ('pwm_method', lazy_gettext(u'PWM Method (Duty Cycle)'))
+        ],
+        validators=[DataRequired()]
+    )
     name = StringField(
         lazy_gettext(u'Name'),
         validators=[DataRequired()]
     )
+    create = SubmitField(lazy_gettext(u'Save'))
+
+
+class TimerAddPWMMethod(FlaskForm):
+    relay_id = StringField(lazy_gettext(u'Relay ID'))
+    method_id = StringField(lazy_gettext(u'Method ID'))
+
+
+class TimerAddRelay(FlaskForm):
     timer_type_output = SelectField(
         lazy_gettext(u'Timer Type'),
         choices=[
@@ -1096,13 +1112,57 @@ class Timer(FlaskForm):
     )
     duration_on = DecimalField(lazy_gettext(u'On (seconds)'))
     duration_off = DecimalField(lazy_gettext(u'Off (seconds)'))
-    create = SubmitField(lazy_gettext(u'Save'))
+
+
+class TimerModBase(FlaskForm):
+    timer_id = IntegerField('Timer ID', widget=widgets.HiddenInput())
+    name = StringField(
+        lazy_gettext(u'Name'),
+        validators=[DataRequired()]
+    )
+    timer_type_output = SelectField(
+        lazy_gettext(u'Timer Type'),
+        choices=[
+            ('relay', lazy_gettext(u'Relay (ON/OFF)')),
+            ('pwm_method', lazy_gettext(u'PWM Method (Duty Cycle)'))
+        ],
+        validators=[DataRequired()]
+    )
     modify = SubmitField(lazy_gettext(u'Save'))
     delete = SubmitField(lazy_gettext(u'Delete'))
     activate = SubmitField(lazy_gettext(u'Activate'))
     deactivate = SubmitField(lazy_gettext(u'Deactivate'))
     order_up = SubmitField(lazy_gettext(u'Up'))
     order_down = SubmitField(lazy_gettext(u'Down'))
+
+
+class TimerModRelay(FlaskForm):
+    relay_id = StringField(lazy_gettext(u'Relay ID'))
+    state = SelectField(
+        lazy_gettext(u'State'),
+        choices=[
+            ('on', lazy_gettext(u'On')),
+            ('off', lazy_gettext(u'Off'))
+        ],
+        validators=[DataRequired()]
+    )
+    time_start = StringField(lazy_gettext(u'Start Time'))
+    time_start_duration = StringField(lazy_gettext(u'Start Time'))
+    time_end_duration = StringField(lazy_gettext(u'End Time'))
+    time_on_duration = DecimalField(
+        lazy_gettext(u'On (seconds)'),
+        validators=[validators.NumberRange(
+            min=0,
+            max=86400
+        )]
+    )
+    duration_on = DecimalField(lazy_gettext(u'On (seconds)'))
+    duration_off = DecimalField(lazy_gettext(u'Off (seconds)'))
+
+
+class TimerModPWMMethod(FlaskForm):
+    relay_id = StringField(lazy_gettext(u'Relay ID'))
+    method_id = StringField(lazy_gettext(u'Method ID'))
 
 
 #